refactor: rename sign out button file
